refactor: improve import performance and fix component naming mismatch
