+++ .github/settings.yml
@@ -1,18 +1,34 @@
 _extends: .github
+
 repository:
   name: cash_lens
   description: Personal finance tracking and analysis built with Elixir and Phoenix.
   topics: elixir, phoenix, finance, tracking, analysis
+  default_branch: master
+  has_issues: true
   has_projects: true
+  has_wiki: false
+  allow_squash_merge: true
+  allow_merge_commit: false
+  allow_rebase_merge: false
+  delete_branch_on_merge: true
 
 branches:
   - name: master
     protection:
+      required_pull_request_reviews:
+        required_approving_review_count: 1
+        dismiss_stale_reviews: true
+        require_code_owner_reviews: true
       required_status_checks:
+        strict: true
         contexts:
           - "GitGuardian Security Checks"
           - "SonarCloud"
           - "SonarCloud Code Analysis"
-          - "DeepSource: Test coverage"
+          - "DeepSource: JavaScript"
           - "DeepSource: Secrets"
           - "Build and test"
+      enforce_admins: true
+      required_conversation_resolution: true
+      restrictions: null
