+++ config/test.exs
@@ -41,3 +41,5 @@ config :phoenix_live_view,
 # Sort query params output of verified routes for robust url comparisons
 config :phoenix,
   sort_verified_routes_query_params: true
+
+config :cash_lens, sql_sandbox: true
